Removed use of /esig and /o from live site
To reduce bandwidth against our (now paid by traffic) website, this version removes use of the /esig/asset redirections and directly points as the GitHub-hosted image assets for the signature block. Also removed the use of the /o/slug redirections and point to the /careers page and the blog tag for aws so we aren't pointing at the now-dead search page.
Also fixed the lint warnings.